+++ src/main/java/exception/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package calculator;
+package exception;
 
 public class ExceptionMessage {
