+++ src/test/java/util/StringUtilsTest.java
@@ -1,6 +1,6 @@
 package util;
 
-import calculator.ExceptionMessage;
+import exception.ExceptionMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
